Moves source files around
