+++ src/app-state/questionnaire/activeQuestionnaire.tsx
@@ -1,6 +1,6 @@
 import { atom } from 'recoil'
 import type { ActiveQuestionnaire } from './ActiveQuestionnaireTypes'
-import { QuestionnaireStateMachine } from '../components/questionnaire/questionnaire-state-machine/QuestionnaireStateMachine'
+import { QuestionnaireStateMachine } from '~/app-state/questionnaire/questionnaire-state-machine/QuestionnaireStateMachine'
 
 const activeQuestionnaire = atom<ActiveQuestionnaire>({
   key: 'activeQuestionnaire',
