+++ src/app-state/questionnaire/questionnaire-state-machine/step-handlers/Connection.tsx
@@ -1,4 +1,4 @@
-import type { StepData } from '~/app-state/ActiveQuestionnaireTypes'
+import type { StepData } from '~/app-state/questionnaire/ActiveQuestionnaireTypes'
 import { v4 as uuidv4 } from 'uuid'
 
 import { Utils } from '../Utils'
