+++ src/to-be-deleted/QuestionnaireStepSelect.tsx
@@ -1,11 +1,11 @@
 import { Select, SelectItem } from '@nextui-org/react'
-import { installationConfig } from '../../app-state/configuration'
+import { installationConfig } from './configuration'
 import type {
   Configuration,
   ConfigurationOption,
   ConfigurationOptionValues,
   ConfigurationOptionsObjects
-} from '../../app-state/configuration'
+} from './configuration'
 import type { Language, QuestionnaireData } from './questionnaireData'
 
 import { useSetRecoilState } from 'recoil'
