+++ src/to-be-deleted/configuration.tsx
